\name{Lacteos}
\alias{Lacteos}
\docType{data}
\title{
Datos reales y resultados de la simulacion de la produccion de quesos madurados
}
\description{
Se muestran los datos reales y los resultados de un modelo de simulacion usando Dinamica de Sistemas bajo el software VENSIM.}
\usage{data(Lacteos)}
\format{
  Un data frame con 12 datos reales y 12 resultados del modelo de simulacion para las siguientes 20 variables, 10 para los datos reales y 10 para la simulacion.
  \describe{
    \item{\code{CompraDeLecheDatos}}{vector numerico}
    \item{\code{VentasDatos}}{vector numerico}
    \item{\code{ProduccionDatos}}{vector numerico}
    \item{\code{DonacionesDatos}}{vector numerico}
    \item{\code{PublicidadDatos}}{vector numerico}
    \item{\code{PerdidasPorDeterioroDatos}}{vector numerico}
    \item{\code{PerdidasPorMaduracionDatos}}{vector numerico}
    \item{\code{TransferenciasDatos}}{vector numerico}
    \item{\code{InventarioGrutaBacterianaDatos}}{vector numerico}
    \item{\code{InventarioCavaVentaDatos}}{vector numerico}
    \item{\code{CompraDeLecheSimula}}{vector numerico}
    \item{\code{VentasSimula}}{vector numerico}
    \item{\code{ProduccionSimula}}{vector numerico}
    \item{\code{DonacionesSimula}}{vector numerico}
    \item{\code{PublicidadSimula}}{vector numerico}
    \item{\code{PerdidasPorDeterioroSimula}}{vector numerico}
    \item{\code{PerdidasPorMaduracionSimula}}{vector numerico}
    \item{\code{TransferenciasSimula}}{vector numerico}
    \item{\code{InventarioGrutaBacterianaSimula}}{vector numerico}
    \item{\code{InventarioCavaVentaSimula}}{vector numerico}
  }
}
\source{
MARQUEZ, Renny. 2007. Un
modelo de simulacion de la
Productora de Alimentos
Universitaria (P.A.U.) Lacteos
Santa Rosa. Proyecto de Grado,
Escuela de Ingenieria de
Sistemas. Merida (Venezuela):
Universidad de Los Andes.
}

\examples{
data(Lacteos)
}
\keyword{datasets}
