\name{Mixteque}
\alias{Mixteque}
\docType{data}
\title{
Dinámica de agua en el suelo durante el ciclo del cultivo de papa en la población de Mixteque en el páramo andino
}
\description{
Datos de un estudio de la dinámica de agua en el suelo durante el ciclo del cultivo de papa en la población de Mixteque en el páramo andino. Se cuenta con tres series de datos para tres posiciones topográficas diferentes (ladera baja, media y alta). Estos datos fueron recolectados por Diaz (2007).
}
\usage{data(Mixteque)}
\format{
  Un data frame con 153 datos para las siguientes 18 variables.
  \describe{
    \item{\code{Mod1SimulaBaja}}{vector numérico}
    \item{\code{Mod1DatosBaja}}{vector numérico}
    \item{\code{Mod1SimulaMedia}}{vector numérico}
    \item{\code{Mod1DatosMedia}}{vector numérico}
    \item{\code{Mod1SimulaAlta}}{vector numérico}
    \item{\code{Mod1DatosAlta}}{vector numérico}
    \item{\code{Mod2SimulaBaja}}{vector numérico}
    \item{\code{Mod2DatosBaja}}{vector numérico}
    \item{\code{Mod2SimulaMedia}}{vector numérico}
    \item{\code{Mod2DatosMedia}}{vector numérico}
    \item{\code{Mod2SimulaAlta}}{vector numérico}
    \item{\code{Mod2DatosAlta}}{vector numérico}
    \item{\code{Mod3SimulaBaja}}{vector numérico}
    \item{\code{Mod3DatosBaja}}{vector numérico}
    \item{\code{Mod3SimulaMedia}}{vector numérico}
    \item{\code{Mod3DatosMedia}}{vector numérico}
    \item{\code{Mod3SimulaAlta}}{vector numérico}
    \item{\code{Mod3DatosAlta}}{vector numérico}
  }
}
\source{
DIAZ, C. 2007. Balance hídrico y de nutrientes y procesos erosivos en un agroecosistema de papa bajo diferentes posiciones topográficas en Los Andes Venezolanos. Avance de tesis doctoral, Facultad de Ciencias, Postgrado en Ecología Tropical. Mérida (Venezuela): Universidad de Los Andes.
}
\examples{
data(Mixteque)
}
\keyword{datasets}
