\name{Lacteos}
\alias{Lacteos}
\docType{data}
\title{
Datos reales y resultados de la simulación de la producción de quesos madurados
}
\description{
Se muestran los datos reales y los resultados de un modelo de simulación usando Dinámica de Sistemas bajo el software VENSIM.}
\usage{data(Lacteos)}
\format{
  Un data frame con 12 datos reales y 12 resultados del modelo de simulación para las siguientes 20 variables, 10 para los datos reales y 10 para la simulación.
  \describe{
    \item{\code{CompraDeLecheDatos}}{vector numérico}
    \item{\code{VentasDatos}}{vector numérico}
    \item{\code{ProduccionDatos}}{vector numérico}
    \item{\code{DonacionesDatos}}{vector numérico}
    \item{\code{PublicidadDatos}}{vector numérico}
    \item{\code{PerdidasPorDeterioroDatos}}{vector numérico}
    \item{\code{PerdidasPorMaduracionDatos}}{vector numérico}
    \item{\code{TransferenciasDatos}}{vector numérico}
    \item{\code{InventarioGrutaBacterianaDatos}}{vector numérico}
    \item{\code{InventarioCavaVentaDatos}}{vector numérico}
    \item{\code{CompraDeLecheSimula}}{vector numérico}
    \item{\code{VentasSimula}}{vector numérico}
    \item{\code{ProduccionSimula}}{vector numérico}
    \item{\code{DonacionesSimula}}{vector numérico}
    \item{\code{PublicidadSimula}}{vector numérico}
    \item{\code{PerdidasPorDeterioroSimula}}{vector numérico}
    \item{\code{PerdidasPorMaduracionSimula}}{vector numérico}
    \item{\code{TransferenciasSimula}}{vector numérico}
    \item{\code{InventarioGrutaBacterianaSimula}}{vector numérico}
    \item{\code{InventarioCavaVentaSimula}}{vector numérico}
  }
}
\source{
MÁRQUEZ, Renny. 2007. Un
modelo de simulación de la
Productora de Alimentos
Universitaria (P.A.U.) Lácteos
Santa Rosa. Proyecto de Grado,
Escuela de Ingeniería de
Sistemas. Mérida (Venezuela):
Universidad de Los Andes.
}

\examples{
data(Lacteos)
}
\keyword{datasets}
