\name{Mixteque}
\alias{Mixteque}
\docType{data}
\title{
Dinamica de agua en el suelo durante el ciclo del cultivo de papa en la poblacion de Mixteque en el paramo andino
}
\description{
Datos de un estudio de la dinamica de agua en el suelo durante el ciclo del cultivo de papa en la poblacion de Mixteque en el paramo andino. Se cuenta con tres series de datos para tres posiciones topograficas diferentes (ladera baja, media y alta). Estos datos fueron recolectados por Diaz (2007).
}
\usage{data(Mixteque)}
\format{
  Un data frame con 153 datos para las siguientes 18 variables.
  \describe{
    \item{\code{Mod1SimulaBaja}}{vector numerico}
    \item{\code{Mod1DatosBaja}}{vector numerico}
    \item{\code{Mod1SimulaMedia}}{vector numerico}
    \item{\code{Mod1DatosMedia}}{vector numerico}
    \item{\code{Mod1SimulaAlta}}{vector numerico}
    \item{\code{Mod1DatosAlta}}{vector numerico}
    \item{\code{Mod2SimulaBaja}}{vector numerico}
    \item{\code{Mod2DatosBaja}}{vector numerico}
    \item{\code{Mod2SimulaMedia}}{vector numerico}
    \item{\code{Mod2DatosMedia}}{vector numerico}
    \item{\code{Mod2SimulaAlta}}{vector numerico}
    \item{\code{Mod2DatosAlta}}{vector numerico}
    \item{\code{Mod3SimulaBaja}}{vector numerico}
    \item{\code{Mod3DatosBaja}}{vector numerico}
    \item{\code{Mod3SimulaMedia}}{vector numerico}
    \item{\code{Mod3DatosMedia}}{vector numerico}
    \item{\code{Mod3SimulaAlta}}{vector numerico}
    \item{\code{Mod3DatosAlta}}{vector numerico}
  }
}
\source{
DIAZ, C. 2007. Balance h?drico y de nutrientes y procesos erosivos en un agroecosistema de papa bajo diferentes posiciones topograficas en Los Andes Venezolanos. Avance de tesis doctoral, Facultad de Ciencias, Postgrado en Ecologia Tropical. Merida (Venezuela): Universidad de Los Andes.
}
\examples{
data(Mixteque)
}
\keyword{datasets}
