\name{PlanktonM}
\alias{PlanktonM}
\docType{data}
\title{
Datos de phytoplankton y biomasa de zooplankton de un modelo}
\description{
Resultados de la simulación de presencia de phytoplankton y biomasa de zooplankton en un lago.
}
\usage{data(PlanktonM)}
\format{
  Un data frame con el resultado de un modelo para 2 variables en 3 instantes de tiempo.
  \describe{
    \item{\code{Instante1Phytoplankton}}{vector numérico}
    \item{\code{Instante2Phytoplankton}}{vector numérico}
    \item{\code{Instante3Phytoplankton}}{vector numérico}
    \item{\code{Instante1Zooplankton}}{vector numérico}
    \item{\code{Instante2Zooplankton}}{vector numérico}
    \item{\code{Instante3Zooplankton}}{vector numérico}
  }
}
\source{
HAEFNER, James W. (2005, pp. 160) Modeling Biological Systems Principles and Applications. Springer.
}

\examples{
data(PlanktonM)
}
\keyword{datasets}
