\name{PlanktonR}
\alias{PlanktonR}
\docType{data}
\title{
Datos observados de fitoplankton y biomasa de zooplankton
}
\description{
Datos observados de la presencia de fitoplankton y biomasa de zooplankton en un lago.
}
\usage{data(PlanktonR)}
\format{
  Un data frame con 6 replicaciones de 2 variables tomadas en 3 instantes de tiempo de cada una.

  \describe{
    \item{\code{Instante1Phytoplankton}}{vector numerico}
    \item{\code{Instante2Phytoplankton}}{vector numerico}
    \item{\code{Instante3Phytoplankton}}{vector numerico}
    \item{\code{Instante1Zooplankton}}{vector numerico}
    \item{\code{Instante2Zooplankton}}{vector numerico}
    \item{\code{Instante3Zooplankton}}{vector numerico}
  }
}

\source{
HAEFNER, James W. (2005, pp. 160) Modeling Biological Systems Principles and Applications. Springer.
}

\examples{
data(PlanktonR)
}
\keyword{datasets}
